BCD finished, Seg7 still working
BCD works for 2 digits and functions correctly with all cases according to testbench. Seg7 has the inputs/outputs mostly finalized and made a possible rough draft of how to format cases with the BCD. NOTE: look for a way to condense since it will be 6 case blocks if done in this method.