Added bcdToSeg7 with testbench and adjusted Seg7
Seg7 is still not completed but added framework and fixed for output length. bcdToSeg7.v and its testbench have been fully completed and successfully takes a BCD value and translates it into ones and tens place 7-Segment display cathode output. This prevents the need for many case statements in the Seg7Display module