Untested Testbench added.
Program current does not compile due to Seg7Display.v having an error: "Error (10161): Verilog HDL error at Seg7Dsiplay.v(182): object "enable_7_seg" is not declared. Verify the object name is correct. If the name is correct, declare the object.". this is due to the enable_7_seg value not being an input or other internal value. Not sure which one it should be since it's a value that needs to be saved between cycles.